Add Debug trait to Storage
