+++ Cargo.toml
@@ -5,9 +5,12 @@ authors = ["Chen Weiguang <chen.weiguang@gmail.com>"]
 
 [dependencies]
 chrono = { version = "0.4", features = ["serde"] }
-derive-getters = "0.0"
+derive-getters = "^0.0.3"
 serde = "1.0"
 serde_derive = "1.0"
+
+[dev-dependencies]
+serde_json = "1.0"
 
 [lib]
 name = "json_collection"
